requirements.txt
asgiref==3.5.2
Django==3.2.13
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.13.0
flake8==4.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
